refactor: Extract folding regions logic to separate file
